broken WIP move minibatch transform and rename/rework
